Music Added to main menu
Need to rework the way music is spawn in the W_menu as for now it's made using spawn 2d sound and sound is promoted to variable /Audio/. The music is stopped during options, going to cutsceen and quiting the game. Works, but needs further improvements.
Menu music is looped from UE options for track.